auth/requirements.txt
redis==4.6.0
fastapi==0.104.1
fastapi-limiter==0.1.5
pydantic==2.3.0
pydantic_settings==2.0.3
orjson==3.9.10
uvicorn==0.23.2
uvloop==0.19.0
wheel==0.41.2
python-dotenv==1.0.0
httptools==0.6.1
SQLAlchemy==2.0.23
alembic==1.12.1
werkzeug==3.0.1
PyJWT==2.8.0
typer==0.9.0
passlib==1.7.4
asyncpg==0.29.0
Authlib==1.3.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0 
opentelemetry-exporter-jaeger==1.21.0
itsdangerous==2.1.2
httpx==0.25.2
psycopg2-binary==2.9.9